Fixes for large velocites along Greenland coast
Greenland runs in CESM on a 4 km grid have been crashing in various situations,
usually with large velocities (> 10 km/yr).  I made several changes:

(1) I fixed a bug in the treatment of velocity at inactive vertices.

The glissade solver computes the velocity only at active vertices, where an
active vertex is any vertex of an active cell, and an active cell is a cell with
thck > thklim. Previously, velocity was not set to zero at inactive vertices.
This meant that if a large velocity was read from an input file at an inactive vertex,
or if the ice retreated from a vertex that previously had a large velocity, the large
velocity could persist indefinitely.  This was usually harmless, but could be
problematic if the large velocity occurs at a vertex with thin ice (0 < thck < thklim)
in neighboring cells.

I simply added a loop to set uvel = vvel = 0 at inactive vertices near the start of the
velocity solve.

(2) I changed the computation of f_ground and beta, introducing a new flotation function
to replace f_pattyn.

The basal traction coefficient, beta, is multiplied by the grounded fraction, f_ground,
so as to zero out the traction under ice shelves and (when the GLP is turned on) to yield
a smooth transition in traction across grounding lines.  Since a GLP was implemented in Mar. 2015,
the grounded fraction has been determined by computing f_pattyn = (-rhow*b)/(rhoi*H)
for each cell and interpolating to vertices.

This procedure is problematic for Greenland fjords, where (for instance) a vertex can be
surrounded by one floating cell with f_pattyn ~ 10 along with 3 well-grounded cells with f_pattyn ~ 0.
As a result, we can have stagf_pattyn > 1 at the vertex, which means in the no-GLP case
that the vertex is ungrounded (f_ground = 0) and beta = 0.  Having beta = 0 at a vertex
lying at the edge of a fjord is dangerous, because large surface gradients can drive very fast flow.

Instead of interpolating f_pattyn, I decided to interpolate 1/f_pattyn = (rhoi*H)/(-rhow*b).
This function goes to infinity as b -> 0 from below, so I capped it at grounding_factor_max = 10
for the case of b > 0 (i.e., above sea level) or b slightly < 0.  I call this new function f_flotation.
Most grounded cells have values >> 1, whereas floating cells have values between 0 and 1.  Thus, one land-based
cell is sufficient to ground a vertex, and 1 or 2 well-grounded marine-based cells can outweigh 2 or 3
floating cells (instead of the reverse, when interpolating f_pattyn).  This reduces the likelihood
of large velocities around coastal fjords.

I still need to verify that for a MISMIP-type problem where we have a more gradual transition
from grounded to floating, there is not much difference between interpolating f_pattyn and f_flotation.
To switch back to the old version, set f_flotation_pattyn = .true. in glissade_masks.F90.

(3) I modified the computation of the basal frictional heat flux, bfricflx.
This flux is given by beta*(u^2 + v^2) at vertices.  Previously, I did a finite-element
calculation, interpolating beta, uvel and vvel within cells and summing over the quadrature points
adjacent to each vertex. A simpler method is just to compute beta*(u^2 + v^2) at vertices.
The FE method is formally more accurate where the field is smooth, but can lead to large errors
in coastal regions where there is a sharp transition between high beta/low vel to low beta/high vel.
In this case, some QPs have both a large beta and a high velocity, leading to excessive basal
friction.  So the code is more robust for Greenland simulations if we revert to the simpler method.

NOTE: The new calcuation changes the answers for standard prognostic tests, including the dome problem.
In order to avoid changes answers in LIVV tests (for now), I introduced a new config option,
which_ho_assemble_bfric.  The old method is option 0 and is the default (for now), while the
new method is option 1.  This is consistent with a couple of other options where a standard FE assembly
method is option 0, and a non-standard option that may be preferable in some situations is option 1.

(4) I fixed a bug in subroutine glissade_grounding_line.  There was a logical error in the
calculation of f_ground in the rare case of two floating cells diagonally opposite a
grounded vertex.

Other changes:

For ease of comparison with CESM, I changed the physical constants shci, lhci and trpt
from parameters to variables that can be specified in the config file.  This allows the
user to insert the standard CESM values in the config file without changing the code.
After making these changes, I verified that the first diagnostic solve for a standalone run
is identical (within roundoff) to a coupled CESM run, provided the input files are the same.
This was a useful sanity check during debugging. Small differences in constants can make a
difference of a few percent in velocities and energy content.

I ran LIVV on a Mac with these changes, and all tests passed BFB.
However, the current tests do not exercise the GLP.